Fixed MCTS to actually sort work
Fixing the skill gain to be faster at lower levels and cap at 40 diff from the skill minimum for the task was the biggest change here.

Our current MCTS is a hack between open and closed, since we recalc the root every planning cycle, but cross our fingers that some of the previously planned nodes might be useful.